# Refactor mapping node traversal and optimize RNode.GetAnnotations and RNode.GetLabels