This PR has two purposes.

First, it refactors mapping node traversal so that all code paths execute through the same function.

Second, it optimizes RNode's GetAnnotations and GetLabels in three ways:

1. For heavily used functions, allocate memory to avoid overhead associated with map and array re-sizing.
2. Where appropriate, limit annotation and label retrievals to only the desired keys.
3. Adjust annotation and label retrieval to avoid unnecessary temporary object creation.

The execution time and memory use improvements over master can be significant. The results shown here are for current master and for the refactored and optimized code. One test case (ds) is a moderately sized and has a somewhat simple set of kustomizations. One (na) is very large and has a complex set of kustomizations. One (ro) is a large set of resources, with no kustomizations - it's just a resource load. kustomize builds using master match exactly with the builds using the PR.

Using go's profiling tools, here's how things look with the current master (a1bfab382) and with this PR (20b0d3c7c):

```
CPU
ds-cpu-mmasterr-a1bfab382.svg: Duration: 12.22s, Total samples = 13.41s (109.72%)
ds-cpu-refactor-20b0d3c7c.svg: Duration: 9.53s, Total samples = 9.12s (95.74%)
na-cpu-mmasterr-a1bfab382.svg: Duration: 186.21s, Total samples = 207.21s (111.28%)
na-cpu-refactor-20b0d3c7c.svg: Duration: 112.41s, Total samples = 110.55s (98.35%)
ro-cpu-mmasterr-a1bfab382.svg: Duration: 44.21s, Total samples = 49.45s (111.86%)
ro-cpu-refactor-20b0d3c7c.svg: Duration: 15.91s, Total samples = 12.80s (80.43%)

Memory
ds-mem-mmasterr-a1bfab382.svg: Showing nodes accounting for 5903.69MB, 76.03% of 7764.73MB total
ds-mem-refactor-20b0d3c7c.svg: Showing nodes accounting for 3108.60MB, 63.65% of 4884.12MB total
na-mem-mmasterr-a1bfab382.svg: Showing nodes accounting for 109331.30MB, 89.11% of 122696.48MB total
na-mem-refactor-20b0d3c7c.svg: Showing nodes accounting for 37047.11MB, 75.00% of 49398.96MB total
ro-mem-mmasterr-a1bfab382.svg: Showing nodes accounting for 26387.54MB, 97.10% of 27174.74MB total
ro-mem-refactor-20b0d3c7c.svg: Showing nodes accounting for 2146.75MB, 82.69% of 2596.03MB total
```

This PR also includes a benchmark test for GetAnnotations. Here's the run against master, followed by the run against the PR:

```
$ go test ./kyaml/yaml/ -bench=. -benchmem -run nope
goos: windows
goarch: amd64
pkg: sigs.k8s.io/kustomize/kyaml/yaml
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
BenchmarkGetAnnotations/00-8             5349039               229.1 ns/op           192 B/op          4 allocs/op
BenchmarkGetAnnotations/02-8             1225515               997.3 ns/op           960 B/op         16 allocs/op
BenchmarkGetAnnotations/05-8              631647              1830 ns/op            1584 B/op         27 allocs/op
BenchmarkGetAnnotations/08-8              463942              2500 ns/op            2016 B/op         36 allocs/op
PASS
ok      sigs.k8s.io/kustomize/kyaml/yaml        6.607s
```

```
$ go test ./kyaml/yaml/ -bench=. -benchmem -run nope
goos: windows
goarch: amd64
pkg: sigs.k8s.io/kustomize/kyaml/yaml
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
BenchmarkGetAnnotations/00-8            12776966                96.90 ns/op           48 B/op          1 allocs/op
BenchmarkGetAnnotations/02-8             4175845               285.9 ns/op           336 B/op          2 allocs/op
BenchmarkGetAnnotations/05-8             3299260               341.5 ns/op           336 B/op          2 allocs/op
BenchmarkGetAnnotations/08-8             2961939               417.2 ns/op           336 B/op          2 allocs/op
PASS
ok      sigs.k8s.io/kustomize/kyaml/yaml        6.544s
```

Closes #4940