@ephesused 2022-12-21
Thanks for the feedback - I hope to have updates (and lint fixes) in the next day or two. I haven't figured out how to get kustomize's linting working properly in my dev environment, so I'm working a bit in the dark on that front.