@ephesused 2022-12-22
Here are the CPU and memory amounts for a0e94c16, the latest commit in the PR. The key point is that, compared to 20b0d3c7c (the prior refactor commit), the changes in a0e94c16 didn't have a negative effect on either run time or memory consumption. (This output also highlights that there's some non-trivial variance between runs on this machine - the master runs previously were 13.41s, 207.21s, and 49.45s.)

```
CPU
ds-cpu-mmasterr-a1bfab382.svg: Duration: 15.83s, Total samples = 15.22s (96.13%)
ds-cpu-refactor-a0e94c164.svg: Duration: 9.54s, Total samples = 9.22s (96.67%)
na-cpu-mmasterr-a1bfab382.svg: Duration: 217.59s, Total samples = 235.22s (108.10%)
na-cpu-refactor-a0e94c164.svg: Duration: 108.78s, Total samples = 109.33s (100.50%)
ro-cpu-mmasterr-a1bfab382.svg: Duration: 52.26s, Total samples = 53.37s (102.13%)
ro-cpu-refactor-a0e94c164.svg: Duration: 13.33s, Total samples = 11.47s (86.05%)

Memory
ds-mem-mmasterr-a1bfab382.svg: Showing nodes accounting for 6074.23MB, 78.28% of 7759.69MB total
ds-mem-refactor-a0e94c164.svg: Showing nodes accounting for 2928.66MB, 61.72% of 4745.28MB total
na-mem-mmasterr-a1bfab382.svg: Showing nodes accounting for 109009.98MB, 89.07% of 122384.84MB total
na-mem-refactor-a0e94c164.svg: Showing nodes accounting for 36909.94MB, 75.09% of 49155.69MB total
ro-mem-mmasterr-a1bfab382.svg: Showing nodes accounting for 26625.33MB, 97.13% of 27412.12MB total
ro-mem-refactor-a0e94c164.svg: Showing nodes accounting for 2169.26MB, 82.00% of 2645.47MB total
```

The benchmark run results for GetAnnotations are pretty much the same as they were with the earlier refactor commit:

```
$ go test ./kyaml/yaml/ -bench=. -benchmem -run nope
goos: windows
goarch: amd64
pkg: sigs.k8s.io/kustomize/kyaml/yaml
cpu: Intel(R) Core(TM) i7-7700 CPU @ 3.60GHz
BenchmarkGetAnnotations/00-8            12347928                89.35 ns/op           48 B/op          1 allocs/op
BenchmarkGetAnnotations/02-8             3991108               284.1 ns/op           336 B/op          2 allocs/op
BenchmarkGetAnnotations/05-8             3191481               343.9 ns/op           336 B/op          2 allocs/op
BenchmarkGetAnnotations/08-8             2733621               421.4 ns/op           336 B/op          2 allocs/op
PASS
ok      sigs.k8s.io/kustomize/kyaml/yaml        7.165s
```